# schema/openapi.json
{
  "components": {
    "schemas": {
...
      "DeletarrDeleteError": {
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path",
          "error"
        ],
        "title": "DeletarrDeleteError",
        "type": "object"
      },
      "DeletarrDeleteRequest": {
        "properties": {
          "paths": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Paths",
            "type": "array"
          },
          "type": {
            "default": "movies",
            "enum": [
              "movies",
              "tv"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "paths"
        ],
        "title": "DeletarrDeleteRequest",
        "type": "object"
      },
      "DeletarrDeleteResult": {
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "integer"
          },
          "deleted_paths": {
            "items": {
              "type": "string"
            },
            "title": "Deleted Paths",
            "type": "array"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/DeletarrDeleteError"
            },
            "title": "Errors",
            "type": "array"
          },
          "failed": {
            "title": "Failed",
            "type": "integer"
          },
          "freed_bytes": {
            "title": "Freed Bytes",
            "type": "integer"
          },
          "freed_formatted": {
            "title": "Freed Formatted",
            "type": "string"
          },
          "freed_mb": {
            "title": "Freed Mb",
            "type": "number"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "deleted",
          "failed",
          "freed_bytes",
          "freed_mb",
          "freed_formatted",
          "deleted_paths",
          "errors"
        ],
        "title": "DeletarrDeleteResult",
        "type": "object"
      },
      "DeletarrLibraryStatus": {
        "properties": {
          "arr_available": {
            "title": "Arr Available",
            "type": "boolean"
          },
          "arr_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arr Detail"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "last_scan_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Scan At"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "results_count": {
            "title": "Results Count",
            "type": "integer"
          },
          "scan_mode": {
            "enum": [
              "heuristic",
              "arr"
            ],
            "title": "Scan Mode",
            "type": "string"
          },
          "stats": {
            "$ref": "#/components/schemas/DeletarrStats"
          },
          "type": {
            "enum": [
              "movies",
              "tv"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "path",
          "last_scan_at",
          "last_error",
          "scan_mode",
          "arr_available",
          "arr_detail",
          "results_count",
          "stats"
        ],
        "title": "DeletarrLibraryStatus",
        "type": "object"
      },
      "DeletarrResults": {
        "properties": {
          "arr_available": {
            "title": "Arr Available",
            "type": "boolean"
          },
          "arr_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arr Detail"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/DeletarrScanItem"
            },
            "title": "Results",
            "type": "array"
          },
          "scan_mode": {
            "enum": [
              "heuristic",
              "arr"
            ],
            "title": "Scan Mode",
            "type": "string"
          },
          "stats": {
            "$ref": "#/components/schemas/DeletarrStats"
          },
          "type": {
            "enum": [
              "movies",
              "tv"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "path",
          "scan_mode",
          "arr_available",
          "arr_detail",
          "results",
          "stats"
        ],
        "title": "DeletarrResults",
        "type": "object"
      },
      "DeletarrScanItem": {
        "properties": {
          "is_checked": {
            "title": "Is Checked",
            "type": "boolean"
          },
          "movie_folder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Movie Folder"
          },
          "movie_folder_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Movie Folder Path"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "origin": {
            "enum": [
              "heuristic",
              "arr"
            ],
            "title": "Origin",
            "type": "string"
          },
          "parent": {
            "title": "Parent",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "type": {
            "enum": [
              "file",
              "folder"
            ],
            "title": "Type",
            "type": "string"
          },
          "videos_in_folder": {
            "items": {
              "$ref": "#/components/schemas/DeletarrVideoRef"
            },
            "title": "Videos In Folder",
            "type": "array"
          }
        },
        "required": [
          "path",
          "name",
          "type",
          "size",
          "reason",
          "parent",
          "movie_folder",
          "movie_folder_path",
          "is_checked",
          "videos_in_folder",
          "origin"
        ],
        "title": "DeletarrScanItem",
        "type": "object"
      },
      "DeletarrScanRequest": {
        "properties": {
          "type": {
            "default": "movies",
            "enum": [
              "movies",
              "tv"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "title": "DeletarrScanRequest",
        "type": "object"
      },
      "DeletarrSettings": {
        "properties": {
          "movies_path": {
            "title": "Movies Path",
            "type": "string"
          },
          "tv_path": {
            "title": "Tv Path",
            "type": "string"
          },
          "use_arr_source": {
            "title": "Use Arr Source",
            "type": "boolean"
          }
        },
        "required": [
          "movies_path",
          "tv_path",
          "use_arr_source"
        ],
        "title": "DeletarrSettings",
        "type": "object"
      },
      "DeletarrSettingsUpdate": {
        "properties": {
          "movies_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Movies Path"
          },
          "tv_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tv Path"
          },
          "use_arr_source": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Use Arr Source"
          }
        },
        "title": "DeletarrSettingsUpdate",
        "type": "object"
      },
      "DeletarrStats": {
        "properties": {
          "is_scanning": {
            "title": "Is Scanning",
            "type": "boolean"
          },
          "scan_progress": {
            "title": "Scan Progress",
            "type": "integer"
          },
          "total_files": {
            "title": "Total Files",
            "type": "integer"
          },
          "total_folders": {
            "title": "Total Folders",
            "type": "integer"
          },
          "total_size": {
            "title": "Total Size",
            "type": "integer"
          }
        },
        "required": [
          "total_files",
          "total_folders",
          "total_size",
          "is_scanning",
          "scan_progress"
        ],
        "title": "DeletarrStats",
        "type": "object"
      },
      "DeletarrStatus": {
        "properties": {
          "libraries": {
            "additionalProperties": {
              "$ref": "#/components/schemas/DeletarrLibraryStatus"
            },
            "propertyNames": {
              "enum": [
                "movies",
                "tv"
              ]
            },
            "title": "Libraries",
            "type": "object"
          },
          "settings": {
            "$ref": "#/components/schemas/DeletarrSettings"
          }
        },
        "required": [
          "settings",
          "libraries"
        ],
        "title": "DeletarrStatus",
        "type": "object"
      },
      "DeletarrVideoRef": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "size"
        ],
        "title": "DeletarrVideoRef",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "All-in-One ARR",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/deletarr/delete": {
      "post": {
        "operationId": "post_delete_api_deletarr_delete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeletarrDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletarrDeleteResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Delete",
        "tags": [
          "deletarr"
        ]
      }
    },
    "/api/deletarr/results": {
      "get": {
        "operationId": "get_results_api_deletarr_results_get",
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "default": "movies",
              "enum": [
                "movies",
                "tv"
              ],
              "title": "Type",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletarrResults"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Results",
        "tags": [
          "deletarr"
        ]
      }
    },
    "/api/deletarr/scan": {
      "post": {
        "operationId": "post_scan_api_deletarr_scan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeletarrScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletarrResults"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Scan",
        "tags": [
          "deletarr"
        ]
      }
    },
    "/api/deletarr/settings": {
      "get": {
        "operationId": "get_settings_api_deletarr_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletarrSettings"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Settings",
        "tags": [
          "deletarr"
        ]
      },
      "put": {
        "operationId": "put_settings_api_deletarr_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeletarrSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletarrStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Settings",
        "tags": [
          "deletarr"
        ]
      }
    },
...
  }
}